(((optimize part19)))
We do not actually need to check with all found scanners combined. It is
sufficient to just look at each original pair of scanners instead of
computing the distances with all already found scanners (compare with
each element in truthl on its own instead of comparing with the whole
truth at once).

This cuts down runtime dramatically.